[Dataset] Generation of chunked dataset
For task with wintermap usage is 10 times faster than the previous version.
Structure of chunked dataset is the following:
```
- data-chunked/
|--chunks/
   |--indexs.npy
   |--y2016/
      |--M01/
         |--d01/
            |--y2016-M01-d01.npy
         |--...
         |--d31/
            |--y2016-M01-d31.npy
      |--M02/
      |--...
      |--M12/
    |--y2017/
    |--...
```
Each file contains 24 hours of data for each day of the month.
Samples concatenated along the channel axis (C, H, W).
Samples can contain different number of channels (with or without winter data).

indexs.npy is a numpy array with shape (n, 7) where n is the number of samples in the dataset.
Each row contains the following information:
- year
- month
- day
- hour
- minute
- start channel
- end channel