[Dataset] Implement skip_withou_wind flag
skip_withou_wind indicates whether to skip the sample if there is no wind data available.
This is useful for the case where we want to train the model with only radar data,
but we have to skip the samples where there is no wind data available.

For example, I have inferenced first stage of the model with both radar and wind data,
but want to train second stage of the model with only radar data.